requirements.txt
pycryptodome==3.20.0
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
Requests==2.32.4
